requirements.txt
apache-airflow==2.5.1
pandas~=2.3.0
requests
dbt-core
dbt-postgres
psycopg2-binary
pandas
openpyxl
requests
python-dotenv~=1.1.1
SQLAlchemy~=1.4.54